+++ app/Http/Controllers/RoleController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreRole;
 use Yajra\DataTables\DataTables;
 use Spatie\Permission\Models\Role;
+use Spatie\Permission\Models\Permission;
 
 class RoleController extends Controller
 {
@@ -16,6 +17,9 @@ class RoleController extends Controller
      */
     public function index()
     {
+        if (!auth()->user()->can('view_role')) {
+            abort(403, 'Unauthorized action');
+        }
         return view('role.index');
     }
 
@@ -24,11 +28,31 @@ class RoleController extends Controller
      */
     public function ssd(Request $request)
     {
+        if (!auth()->user()->can('view_role')) {
+            abort(403, 'Unauthorized action');
+        }
         $role = Role::query();
         return DataTables::of($role)
+            ->addColumn('permissions', function ($each) {
+                $output = '';
+                foreach ($each->permissions as $permission) {
+                    $output .= '<span class="badge badge-pill badge-primary m-1">' . $permission->name . '</span>';
+                }
+                return $output;
+            })
             ->addColumn('action', function ($each) {
-                $edit_icon = '<a href="' . route('role.edit', $each->id) . '" class="text-warning"><i class="fas fa-edit"></i></a>';
-                $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="' . $each->id . '"><i class="fas fa-trash-alt"></i></a>';
+                $edit_icon = '';
+                $delete_icon = '';
+
+                if (!auth()->user()->can('edit_role')) {
+                    $edit_icon = '<a href="' . route('role.edit', $each->id) . '" class="text-warning"><i class="fas fa-edit"></i></a>';
+                }
+                if (!auth()->user()->can('delete_role')) {
+
+                    $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="' . $each->id . '"><i class="fas fa-trash-alt"></i></a>';
+                }
+
+
 
                 return '<div class="action-icon">'  . $edit_icon . $delete_icon . '</div>';
             })
@@ -38,12 +62,16 @@ class RoleController extends Controller
             ->addColumn('plus_icon', function ($each) {
                 return null;
             })
-            ->rawColumns(['action'])
+            ->rawColumns(['action', 'permissions'])
             ->make(true);
     }
     public function create()
     {
-        return view('role.create');
+        if (!auth()->user()->can('create_role')) {
+            abort(403, 'Unauthorized action');
+        }
+        $permissions = Permission::all();
+        return view('role.create', compact('permissions'));
     }
 
     /**
@@ -51,11 +79,15 @@ class RoleController extends Controller
      */
     public function store(StoreRole $request)
     {
+        if (!auth()->user()->can('create_role')) {
+            abort(403, 'Unauthorized action');
+        }
         $role = new Role;
         $role->name = $request->name;
         $role->save();
+        $role->givePermissionTo($request->permissions);
 
-        return redirect()->route('role.index')->with('create','Successfully Create Role');
+        return redirect()->route('role.index')->with('create', 'Successfully Create Role');
     }
 
     /**
@@ -71,20 +103,29 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
+        if (!auth()->user()->can('edit_role')) {
+            abort(403, 'Unauthorized action');
+        }
         $role = Role::findOrFail($id);
-        return view('role.edit',compact('role'));
+        $old_permissions = $role->permissions->pluck('id')->toArray();
+        $permissions = Permission::all();
+        return view('role.edit', compact('role', 'permissions', 'old_permissions'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateRole $request,$id)
+    public function update(UpdateRole $request, $id)
     {
+        if (!auth()->user()->can('edit_role')) {
+            abort(403, 'Unauthorized action');
+        }
         $role = Role::findOrFail($id);
         $role->name = $request->name;
         $role->update();
 
-        return redirect()->route('role.index')->with('edit','Successfully Edit Role');
+        $role->syncPermissions($request->permissions);
+        return redirect()->route('role.index')->with('edit', 'Successfully Edit Role');
     }
 
     /**
@@ -92,6 +133,9 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        if (!auth()->user()->can('delete_role')) {
+            abort(403, 'Unauthorized action');
+        }
         $role = Role::findOrFail($id);
         $role->delete();
         return 'success';
